Fix remove bug
