+++ src/modules/interractions.js
@@ -23,27 +23,10 @@ export const trueFalse = () => {
   });
 };
 
-const clearCompleted = document.querySelector('#clear-completed');
-
-const del = () => {
+export const del = () => {
   Array.from(document.querySelectorAll('.checkBox')).forEach((complete) => {
     if (complete.checked) {
       complete.parentElement.remove();
     }
-  });
-};
-
-export const deleteItems = () => {
-  clearCompleted.addEventListener('click', () => {
-    const filterAll = todos.filter((item) => item.completed !== true);
-    todos.length = 0;
-    filterAll.forEach((item) => {
-      todos.push(item);
-    });
-    todos.forEach((todo, id) => {
-      todo.id = id;
-    });
-    localStorage.setItem('todos', JSON.stringify(todos));
-    del();
   });
 };
